docker/preprocessing/requirements.txt
kaggle==1.6.17
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
Pillow==10.4.0
Requests==2.32.3
schedule==1.2.2
uvicorn==0.30.6
tqdm==4.66.5